requirements.txt
python-telegram-bot==13.15
APScheduler==3.6.3
Flask==2.2.5
requests==2.32.3
python-dotenv==1.0.1
gunicorn==20.1.0
urllib3==1.26.15
